Implemented randomized condition order. - First working version.
